+++ src/main/java/com/batchproject/jobs/models/housing/SuiteRepository.java
@@ -2,5 +2,5 @@ package com.batchproject.jobs.models.housing;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SuiteRepository extends JpaRepository<Suite, Integer> {
+public interface SuiteRepository extends JpaRepository<Suite, Long> {
 }
