+++ src/main/java/com/batchproject/jobs/models/maintenance/WorkOrderDTO.java
@@ -11,7 +11,7 @@ public class WorkOrderDTO {
     private LocalDate completedDate;
     private String status; // e.g., "Pending", "In Progress", "Completed"
     private String notes;
-    private Integer suiteId;       // ID of the associated Suite
-    private Integer buildingId;    // ID of the associated HousingBuilding
-    private Integer maintenanceId;
+    private Long suiteId;       // ID of the associated Suite
+    private Long buildingId;    // ID of the associated HousingBuilding
+    private Long maintenanceId;
 }
