+++ src/main/java/com/batchproject/jobs/models/maintenance/WorkOrderRepository.java
@@ -2,5 +2,5 @@ package com.batchproject.jobs.models.maintenance;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WorkOrderRepository extends JpaRepository<WorkOrder, Integer> {
+public interface WorkOrderRepository extends JpaRepository<WorkOrder, Long> {
 }
